another cleanup push

+++ 573Goals/Models/Goal.swift
@@ -12,19 +12,19 @@ import CoreData
 @objc(GoalEntity)
 public class GoalEntity: NSManagedObject {
     
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Double
     @NSManaged public var month: String
-    @NSManaged public var amount: Int64
+    @NSManaged public var amount: Double
     @NSManaged public var name: String
     @NSManaged public var percentage: Float
     
     
-    class func createInManagedObjectContext(_ context: NSManagedObjectContext, id: Int64, month: String, amount: Int64, name: String, percentage: Float) -> GoalEntity {
+    class func createInManagedObjectContext(_ context: NSManagedObjectContext, id: Double, month: String, amount: Double, name: String, percentage: Float) -> GoalEntity {
          let newGoalEntity = NSEntityDescription.insertNewObject(forEntityName: "GoalEntity", into: context) as! GoalEntity
             newGoalEntity.id = id
             newGoalEntity.month = month
             newGoalEntity.name = name
-            newGoalEntity.amount = Int64(amount) 
+            newGoalEntity.amount = Double(amount)
             newGoalEntity.percentage = percentage
             
             return newGoalEntity
